Rename Login to Login; xml
This code is an XML layout file for an Android app that creates a login screen. The file contains a RelativeLayout with several child views inside it. Here is a breakdown of each child view:

TextView: This view has an ID of "login_title" and displays the text "Login". It is centered horizontally and has a top margin of 50dp.
EditText: This view has an ID of "email_edittext" and displays a hint of "Email". It has an input type of "textEmailAddress" and is positioned below the login_title TextView. It also has a top margin of 100dp and left and right margins of 20dp.
EditText: This view has an ID of "password_edittext" and displays a hint of "Password". It has an input type of "textPassword" and is positioned below the email_edittext EditText. It also has a top margin of 20dp and left and right margins of 20dp.
Button: This view has an ID of "login_button" and displays the text "Login". It is positioned below the password_edittext EditText and has a top margin of 20dp. It also has left and right margins of 20dp.
TextView: This view has an ID of "register_link" and displays the text "Not registered? Register here". It is positioned below the login_button Button and has a top margin of 20dp. It is centered horizontally and has a text size of 16sp.
Overall, this XML layout file creates a simple login screen with email and password input fields, a login button, and a link to register.